feat(wren-ui): rename page, path, move api reference link to sidebar
