+++ wren-ui/src/components/sidebar/index.tsx
@@ -8,7 +8,7 @@ import SettingOutlined from '@ant-design/icons/SettingOutlined';
 import Home, { Props as HomeSidebarProps } from './Home';
 import Modeling, { Props as ModelingSidebarProps } from './Modeling';
 import Knowledge from './Knowledge';
-import APIHistory from './APIHistory';
+import APIManagement from './APIManagement';
 import LearningSection from '@/components/learning';
 
 const Layout = styled.div`
@@ -64,8 +64,8 @@ const DynamicSidebar = (
       return <Knowledge />;
     }
 
-    if (pathname.startsWith(Path.APIHistory)) {
-      return <APIHistory />;
+    if (pathname.startsWith(Path.APIManagement)) {
+      return <APIManagement />;
     }
 
     return null;
