+++ wren-ui/src/utils/enum/menu.ts
@@ -2,4 +2,5 @@ export enum MENU_KEY {
   QUESTION_SQL_PAIRS = 'question-sql-pairs',
   INSTRUCTIONS = 'instructions',
   API_HISTORY = 'api-history',
+  API_REFERENCE = 'api-reference',
 }
